feat: run selector for fallback handler options too
If handlers are found in the "Added Associations" or system
applications, then also filter them and run the selector if necessary,
just like it would be done for the "Default Associations".
