+++ src/config/mod.rs
@@ -1,5 +1,4 @@
 mod config_file;
 mod main_config;
 
-pub use config_file::ConfigFile;
 pub use main_config::{get_languages, Config, Languages};
